added more folders to components
