+++ client/components/index.js
@@ -5,13 +5,13 @@
  */
 
 // export {default as Home} from './home'
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
+export {default as NavBar} from './navbar'
+export {default as UserHome} from './users/user-home'
 export {Login, Signup} from './auth-form'
-export {AllProducts} from './all-products'
-export {default as SingleProduct} from './SingleProduct'
-export {default as AllUsers} from './allusers'
-export {default as AdminHome} from './admin-home'
+export {AllProducts} from './products/all-products'
+export {default as SingleProduct} from './products/singleproduct'
+export {default as AllUsers} from './users/allusers'
+export {default as AdminHome} from './admin_routes/admin-home'
 export {default as Home} from './home'
 
 export {ErrorPage} from './error-page'
