+++ client/components/products/all-products.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import PropTypes from 'prop-types'
-import {fetchProducts, removeProduct} from '../store/allProducts'
+import {fetchProducts, removeProduct} from '../../store/allProducts'
 import {BrowserRouter as Router, Link, withRouter} from 'react-router-dom'
 import Loader from 'react-loader-spinner'
 
@@ -26,7 +25,6 @@ class AllProductsDC extends React.Component {
     this.setState({selection: event.target.value})
   }
 
-
   handleDelete(id) {
     this.props.deleteProduct(id)
   }
@@ -35,7 +33,6 @@ class AllProductsDC extends React.Component {
     const loading = this.props.loading
     const allProducts = this.props.allProducts || []
     const role = this.props.role || 'User'
-
 
     //filtering products
     //need to refactor once DB gets refactored...
@@ -85,7 +82,6 @@ class AllProductsDC extends React.Component {
                 </a>
                 <Link to={`/singleproduct/${elm.id}`}>{elm.name}</Link>
 
-
                 {/* If the user is an admin, button to delete product */}
                 {role === 'Admin' ? (
                   <button
@@ -97,7 +93,6 @@ class AllProductsDC extends React.Component {
                 ) : (
                   ''
                 )}
-
               </div>
             )
           })}
