+++ client/components/users/allusers.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import {fetchUsers} from '../store/admin'
+import {fetchUsers} from '../../store/admin'
 
 import './allusers.css'
 
-class allusers extends React.Component {
+class AllUsersDC extends React.Component {
   constructor(props) {
     super(props)
     this.state = {}
@@ -57,4 +57,4 @@ const mapDispatch = dispatch => {
   }
 }
 
-export default connect(mapState, mapDispatch)(allusers)
+export default connect(mapState, mapDispatch)(AllUsersDC)
